Disable minification (#57)
Fix the hot-reloading for the javascript sources by disabling the minification of js files.

The minification isn't super effective anyway and is done through a custom script instead of using a generic bundler like webpack. I would recommend to stick without it until we start to develop the new frontend.